# lbaw2124/openapi.yaml
openapi: '3.0.2'
info:
  title: FeupTech
  version: '1.0'
  description: 'Web Resources Specification (A7) for FeupTech'
servers:
  - url: http://lbaw2224.lbaw.fe.up.pt
    description: 'Production Server'
...
paths:
...
  /login:
    get:
      operationId: R101
      summary: 'R101: Login Form'
      description: 'Provide login form. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
        '200':
          description: 'Ok. Show Log-in UI'
    post:
      operationId: R102
      summary: 'R102: Login Action'
      description: 'Process the login form submission. Access: PUB'
      tags:
        - 'M01: Authentication and Individual Profile'

      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                email:       # <!--- form field name
                  type: string
                  format: email
                password:    # <!--- form field name
                  type: string
                  format: password
              required:
                    - email
                    - password

      responses:
        '302':
          description: 'Redirect after processing the login credentials.'
          content:
            application/json:
              schema:
                type: string
              examples:
                Success:
                  description: 'Successful authentication. Redirect to user profile.'
                  value: '/personalinfo'
                Error:
                  description: 'Failed authentication. Redirect to login form.'
                  value: '/login'
...
  /personalinfo:
     get:
      operationId: R108
      summary: 'R108: View User Profile'
      description: 'Show the individual user profile. Access: USR'
      tags:
        - 'M01: Authentication and Individual Profile'

      responses:
        '200':
          description: 'Ok. Show User Profile UI'
     post:
      operationId: R109
      summary: 'R109: Appeal Unblock'
      description: 'Appeal a "Block" in order to be able to make reviews again. Access: USR'
      tags:
        - 'M01: Authentication and Individual Profile'
      responses:
          '200':
            description: 'Ok. Request in queue.'

     delete:
      operationId: R109
      summary: 'R109: Delete User Profile'
      description: 'Delete the individual user profile. Access: USR'
      tags:
        - 'M01: Authentication and Individual Profile'
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                current password:
                  type: string
                  format: password
              required:
                    - current password
      responses:
          '302':
            description: 'Evaluate if account should be deleted.'
            content:
                  application/json:
                    examples:
                      Success:
                        description: 'Password accepted. Redirect to /products.'
                      Failure:
                        description: 'Password rejected. Redirect to /personainfo.'
          '404': 
            description: 'Profile not found.'
     patch:
      operationId: R110
      summary: 'R110: Edit User Profile Request'
      description: 'File a profile edit request. Access: USR'
      tags: 
        - 'M01: Authentication and Individual Profile'
      requestBody:
            required: true
            content:
              application/x-www-form-urlencoded:
                schema:
                  type: object
                  properties:
                    new picture:
                      type: string
                      format: binary
                    new name:
                      type: string
                    new email:
                      type: string
                      format: email
                    new password:
                      type: string
                      format: password
                    confirm new password:
                      type: string
                      format: password
                    current password:       # <!--- form field name
                      type: string
                      format: password
                  required:
                        - current password
      responses:
              '200':
                description: 'Redirect after processing the new information.'
                content:
                  application/json:
                    examples:
                      Success:
                        description: 'Update.'
                        
                      Failure:
                        description: 'Incorrect current password. Do not update'
...